fix(gui): resolve shift+hover highlighting bug
Fixed critical bug where objects would not highlight when holding shift
and hovering over them.

Root cause: When refactoring from signed to unsigned integers, the
expression `max(0U, ux - looksize)` in get_near_item() and get_near_char()
would underflow when ux < looksize. This caused the subtraction to wrap
around to ~4 billion, making the start index larger than the end index,
so the search loop never executed and no objects were ever detected.

Example: ux=25, looksize=51 -> (25-51) wraps to 4294967270
         max(0U, 4294967270) = 4294967270, causing sx > ex

Changes:
- Replace max(0U, ux - looksize) with ternary check to prevent underflow
- Change context_getnm() return type from int to size_t with MAXMN sentinel
- Fix type consistency: cast mn to size_t (not unsigned int) when comparing
  with itmsel/chrsel which are size_t
- Change nearest variable from unsigned int to size_t in get_near_* functions

+++ src/gui/gui.h
@@ -63,7 +63,7 @@ DLL_EXPORT int _get_skltab_index(int n);
 extern int (*get_skltab_show)(int i);
 DLL_EXPORT int _get_skltab_show(int i);
 
-int context_getnm(void);
+size_t context_getnm(void);
 int context_action_enabled(void);
 int hover_capture_text(char *line);
 void hover_capture_tick(void);
